Fix bug with end of ranges
The initial implementation included all the lines to the next
declaration as the "end" of of a declaration. However, with
multiple files and headers, the next declaration could be in
a different file.

Only allowing the next declaration in the same file exposed a bug
with the casting used in FindRBrace, because of a weird interaction
between dyn_cast and multiple inheritance.

Fixing that bug showed that the default location of the "end of main
file" was now being used, which is also silly since the default
constructor for clang::SourceLocation just produces a dummy
(and checkably) invalid value.

Replacing that showed that the FindRBrace vs decl_end code was silly,
since presumbly getBraceRange()->getEnd() == getEndLoc() (checked
against tests) and so could be deleted.

Similarly, the until-next-decl code was silly too - deleting that showed
another casting bug where a DeclContext* was being passed to a Decl*
recursive call.

At this point in deleting, all the code just seemed to recurse or
fancily call getEndLoc() one way or another.

This took a stupid amount of time to diagnose and fix, especially
considering the fix was to just delete all the complicated code.