Fix bug + tidy up code
This commit fixes a bug. Previously, macro dependencies were
erroneously ignored if they were not recorded as being marked.
Now, macro dependencies are marked if they are not recorded as being
marked.

Otherwise, all debug output is under the SIMP_DEBUG macro, and the
code is a little better commented.