Fix bug with macros with lines starting w/ strings
And add tests to cover this. Basically, the source range
of a macro ends at the _beginning_ of the last token, rather
than at the end of it. Further, there's some strange interaction
with the length of a string literal in a macro. Anyways, this
works around both of these issues.