Fix bug: validate compile commands input
One file can have multiple commands, and the FrontendAction gets run for
each, resulting in a big union of marked lines. This checks that a
command is unique for the given file.

Option parser is probably unnecessarily fancy.

SingletonDb is there because FixedCompilationDatabase futzes with the
command too much and thus couldn't be made to work.